requirements.txt
flask
pandas
numpy
geopy
pydantic>=2.0
openai
python-dotenv
supabase==2.0.0
